patch

+++ ts/updateExistingTagComponent.function.ts
@@ -2,9 +2,10 @@ import { TagSubject } from './Tag.utils'
 import { TemplaterResult } from './TemplaterResult.class'
 import { Tag } from './Tag.class'
 import { hasTagSupportChanged } from './hasTagSupportChanged.function'
-import { destroyTagMemory } from './checkDestroyPrevious.function'
 import { TagSupport, renderTagSupport } from './TagSupport.class'
 import { processSubjectComponent } from './processSubjectComponent.function'
+import { destroyTagMemory } from './destroyTag.function'
+import { deepClone } from './deepFunctions'
 import { State } from './set.function'
 
 export function updateExistingTagComponent(
@@ -14,6 +15,13 @@ export function updateExistingTagComponent(
   insertBefore: Element | Text,
 ): void {
   let existingTag: Tag | undefined = subject.tag
+  
+  /*
+  if(existingTag && !existingTag.hasLiveElements) {
+    throw new Error('issue already began')
+  }
+  */
+  
   const oldWrapper = existingTag.tagSupport.templater.wrapper
   const newWrapper = tempResult.wrapper
   let isSameTag = false
@@ -45,18 +53,6 @@ export function updateExistingTagComponent(
     })
     return
   } else {
-    // const subjectTagSupport = existingTag.tagSupport
-    // old props may have changed, reclone first
-    
-    /*
-    const newProps = subjectTagSupport.propsConfig.latest
-    let oldCloneProps = subjectTagSupport.propsConfig.clonedProps
-    // if the new props are NOT HTML children, then clone the props for later render cycle comparing
-    if(!isTagInstance(newProps)) {
-      oldCloneProps = deepClone( newProps )
-    }
-    */
-
     if(!tempResult.tagSupport) {
       tempResult.tagSupport = new TagSupport(
         oldTagSupport.ownerTagSupport,
@@ -65,77 +61,40 @@ export function updateExistingTagComponent(
       )
     }
 
-
     const newTagSupport = tempResult.tagSupport
-
-    /*
-    const hasStateChanged = checkStateChanged(oldTagSupport.memory.state)
-    if(hasStateChanged) {
-      console.log('support changed!!!!')
-    }
-    */
-    // const hasChanged = hasStateChanged || hasTagSupportChanged(oldTagSupport, newTagSupport, tempResult)
     const hasChanged = hasTagSupportChanged(oldTagSupport, newTagSupport, tempResult)
+
     if(!hasChanged) {
-      console.log('!hasChanged', {
-        original: tempResult.wrapper.original,
-        old: oldTagSupport.propsConfig.latest,
-        new: newTagSupport.propsConfig.latest,
-        tempProps: tempResult.props,
-        equal: oldTagSupport.propsConfig.latest == newTagSupport.propsConfig.latest,
-        state: newTagSupport.memory.state.newest,
-        oldState: oldTagSupport.memory.state.newest,
-      })
+      const stateChanged = false
+      if(stateChanged) {
+        const newTag = renderTagSupport(
+          tempResult.tagSupport,
+          false,
+        )
+  
+        const oldTag = oldGlobal.oldest as Tag
+        oldTag.updateByTag(newTag)
+      }
 
-      const newTag = renderTagSupport(
-        tempResult.tagSupport,
-        false,
-      )
-
-      const oldTag = oldGlobal.oldest as Tag
-      oldTag.updateByTag(newTag)
-    
       return // its the same tag component
     }
   }
 
-  // setValueRedraw(tempResult, subject, ownerTag)
-
-  // ???
-  // oldTagSupport.templater = tempResult
-    
   const oldestTag = tempResult.global.oldest as Tag // oldTagSupport.oldest as Tag // existingTag
   const previous = tempResult.global.newest as Tag
 
   if(!previous || !oldestTag) {
-    throw new Error('how?')
+    throw new Error('how no previous or oldest nor newest?')
   }
 
-  console.log('updateExistingTagComponent.function.ts - start', {
-    original: tempResult.wrapper.original,
-    props: tempResult.props,
-    oldProps: subject.tag.tagSupport.templater.props,
-    oldestProps: subject.tag.tagSupport.templater.global.oldest?.tagSupport.templater.props,
-    
-    previousState: previous.tagSupport.memory.state.newest,
-    newestState: tempResult.tagSupport.memory.state.newest,
-    olderState: tempResult.tagSupport.subject.tag.tagSupport.memory.state.newest,
-  })
   const newTag = renderTagSupport(
     tempResult.tagSupport,
     false,
   )
 
+  existingTag = subject.tag
+
   const hasOldest = newTag.tagSupport.templater.global.oldest ? true : false
-
-  console.log('updateExistingTagComponent.function.ts - done', {
-    original: tempResult.wrapper.original,
-    props: tempResult.props,
-    oldProps: subject.tag.tagSupport.templater.props,
-    oldestProps: subject.tag.tagSupport.templater.global.oldest?.tagSupport.templater.props,
-    hasOldest,
-  })
-
   if(!hasOldest) {
     newTag.buildBeforeElement(oldInsertBefore, {
       forceElement: true,
@@ -153,9 +112,9 @@ export function updateExistingTagComponent(
 
     subject.tag = newTag
 
-    oldTagSupport.templater.global.newest = newTag
-    // ???
-    // oldTagSupport.propsConfig = {...tempResult.tagSupport.propsConfig}
+    if(!newTag.hasLiveElements) {
+      throw new Error('44444 - 5')
+    }  
   
     return
   }
@@ -173,48 +132,44 @@ export function updateExistingTagComponent(
     throw new Error('bad elders')
   }
 
+  let oldest = oldTagSupport.templater.global.oldest
+
   if(isLikeTag) {
     if(!newTag.tagSupport.templater.global.oldest) {
       throw new Error('maybe 6')
     }
 
     subject.tag = newTag
-    // existingTag.updateByTag(newTag)
+    /*
+    if(!newTag.hasLiveElements) {
+      throw new Error('44444 - 6')
+    }
+    */
+
     oldestTag.updateByTag(newTag) // the oldest tag has element references
   } else {
     // Although function looked the same it returned a different html result
-    if(isSameTag) {
+    if(isSameTag && existingTag) {
       destroyTagMemory(existingTag, subject)
+      newTag.tagSupport.templater.global.context = {} // do not share previous outputs
     }
+    oldest = undefined
 
-    // oldTagSupport.oldest = newTag
-    if(!newTag.tagSupport.templater.global.oldest) {
-      throw new Error('maybe 7')
-    }
-
-    subject.tag = newTag
+    // ??? - new remove
+    // subject.tag = newTag
   }
 
-  if(!oldTagSupport.templater.global.oldest) {
+  if(!oldest) {
     newTag.buildBeforeElement(oldTagSupport.templater.global.insertBefore as Element, {
       forceElement: true,
       counts: {added: 0, removed: 0}, test:false,
     })
 
     newTag.tagSupport.templater.global.oldest = newTag
-    oldTagSupport.templater.global.oldest = newTag
     newTag.tagSupport.templater.global.newest = newTag
-    oldTagSupport.templater.global.newest = newTag
-
-    if(!newTag.tagSupport.templater.global.oldest) {
-      throw new Error('maybe 8')
-    }
+    oldTagSupport.templater.global.oldest = newTag
 
     subject.tag = newTag
-
-    if(!newTag.hasLiveElements) {
-      throw new Error('55555')
-    }
   }
 
   oldTagSupport.templater.global.newest = newTag
@@ -234,6 +189,8 @@ function checkStateChanged(state: State) {
       return true
     }
 
+    console.log('not matched', {lastValue, nowValue})
     return false
   })
 }
+
